+++ database/migrations/2020_05_21_300000_create_team_invitations_table.php
@@ -16,9 +16,11 @@ class CreateTeamInvitationsTable extends Migration
         Schema::create('team_invitations', function (Blueprint $table) {
             $table->id();
             $table->foreignId('team_id')->constrained()->cascadeOnDelete();
-            $table->string('email')->unique();
+            $table->string('email');
             $table->string('role')->nullable();
             $table->timestamps();
+
+            $table->unique(['team_id', 'email']);
         });
     }